Refactor Pokerboy.GameServer
This is a refactor which attempts to remove the bulk of the logic
out of the gameserver.  Hopefully further refactors will clean this
up a bit more and make adding more functionality a bit more testable.

Another item this pulls in is dialyzer; which will hopefully find type errors
and prevent errors from getting into a build.