Update Pokerboy to Latest Elixir / Phoenix (#1)
* Update to latest Phoenix / Elixir

* Make votes configureable

* Add Github Tag to Demo

* Add in Credo and fix all issues found by it

* Rename to game_channel to match module

* Change file structure to match modern phoenix app

* Refactor GameServer to match normal naming

* Rename GameSupervisor to match normal naming

* Refactor Pokerboy.GameServer

This is a refactor which attempts to remove the bulk of the logic
out of the gameserver.  Hopefully further refactors will clean this
up a bit more and make adding more functionality a bit more testable.

Another item this pulls in is dialyzer; which will hopefully find type errors
and prevent errors from getting into a build.

* Testing the state changing functions for Game (#2)

Before introducting more moving parts I want to cover what I have
in tests so the rug doesn't fly out from under me.